[Refactor] 결과 출력 processor ReturnTotalProcessor.kt -> Lotto.kt 변경
변경에 따른 LottoGame.kt 속 파일명 변경
